requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
openai>=1.12.0
google-generativeai>=0.3.0
anthropic>=0.8.1
graphviz>=0.20.1
requests>=2.31.0 